Botania
Ported from 1.10. Lots of additions and changes, including tall flowers
made from the alt textures, so I put them as the default again with
some edits to some of them. Had to make new textures for flowers 11 to
15, which ended up pretty derivative, some being edited Faithful
textures, and a couple inspired by Soartex. There are still a lot of
improvements to be made on the flowers, but I think they're more
consistent now overall.